removed previous monitoring implemenatation
