+++ src/app.module.ts
@@ -12,7 +12,6 @@ import { AdminController } from './admin/admin.controller';
 import { GetConfig } from './config/config';
 import { SeedService } from './shared/seed/seed.service';
 import { MonitoringModule } from './monitoring/monitoring.module';
-import { MonitoringModuleNew } from './monitoring-new/monitoring.module';
 
 @Module({
   imports: [
@@ -22,7 +21,6 @@ import { MonitoringModuleNew } from './monitoring-new/monitoring.module';
     PaymentModule,
     UserModule,
     MonitoringModule,
-    MonitoringModuleNew,
   ],
   controllers: [AppController, StatisticController, AdminController],
   providers: [StatisticService, CfpService, SeedService],
