+++ src/monitoring/monitoring.controller.ts
@@ -1,9 +1,10 @@
-import { Controller, Get, UseGuards } from '@nestjs/common';
+import { Body, Controller, Get, Post, Query, UseGuards } from '@nestjs/common';
 import { AuthGuard } from '@nestjs/passport';
 import { ApiBearerAuth, ApiExcludeEndpoint, ApiTags } from '@nestjs/swagger';
 import { RoleGuard } from 'src/shared/auth/role.guard';
 import { UserRole } from 'src/shared/auth/user-role.enum';
 import { MonitoringService } from './monitoring.service';
+import { SystemState, SubsystemState, Metric } from './system-state-snapshot.entity';
 
 @ApiTags('monitoring')
 @Controller('monitoring')
@@ -14,12 +15,22 @@ export class MonitoringController {
   @ApiBearerAuth()
   @ApiExcludeEndpoint()
   @UseGuards(AuthGuard(), new RoleGuard(UserRole.ADMIN))
-  async getMonitoringData(): Promise<any> {
-    return {
-      user: await this.monitoringService.getUser(),
-      staking: await this.monitoringService.getStaking(),
-      payment: await this.monitoringService.getPayment(),
-      node: await this.monitoringService.getNode(),
-    };
+  async getSystemState(
+    @Query('subsystem') subsystem: string,
+    @Query('metric') metric: string,
+  ): Promise<SystemState | SubsystemState | Metric> {
+    return await this.monitoringService.getState(subsystem, metric);
+  }
+
+  @Post('data')
+  @ApiBearerAuth()
+  @ApiExcludeEndpoint()
+  @UseGuards(AuthGuard(), new RoleGuard(UserRole.ADMIN))
+  async onWebhook(
+    @Query('subsystem') subsystem: string,
+    @Query('metric') metric: string,
+    @Body() data: unknown,
+  ): Promise<any> {
+    return await this.monitoringService.onWebhook(subsystem, metric, data);
   }
 }
