+++ src/monitoring/observers/staking-balance.observer.ts
@@ -2,8 +2,8 @@ import { Injectable } from '@nestjs/common';
 import { Interval } from '@nestjs/schedule';
 import { WhaleService } from 'src/ain/whale/whale.service';
 import { Config } from 'src/config/config';
-import { MetricObserver } from 'src/monitoring-new/metric.observer';
-import { MonitoringService } from 'src/monitoring-new/monitoring.service';
+import { MetricObserver } from 'src/monitoring/metric.observer';
+import { MonitoringService } from 'src/monitoring/monitoring.service';
 import { CryptoInput } from 'src/payment/models/crypto-input/crypto-input.entity';
 import { CryptoStakingRepository } from 'src/payment/models/crypto-staking/crypto-staking.repository';
 import { MasternodeRepository } from 'src/payment/models/masternode/masternode.repository';
