+++ src/monitoring/system-state-snapshot.entity.ts
