+++ src/monitoring/system-state-snapshot.repository.ts
