+++ thunder-tests/thunderclient.json
@@ -121,7 +121,7 @@
     "method": "POST",
     "sortNum": 80000,
     "created": "2022-01-15T13:41:33.269Z",
-    "modified": "2022-04-07T11:56:26.698Z",
+    "modified": "2022-07-05T06:47:40.995Z",
     "headers": [],
     "params": [],
     "body": {
@@ -866,13 +866,28 @@
     "colId": "62eefab7-84cd-496b-8f93-253c15d4f841",
     "containerId": "fd2e8b01-0e57-4f22-bbb5-226617dc19db",
     "name": "Get Monitoring Data",
-    "url": "{{url}}/v1/monitoring-new/data",
+    "url": "{{url}}/v1/monitoring/data?subsystem=bankingBot&metric=logs",
     "method": "GET",
     "sortNum": 330000,
     "created": "2022-04-12T13:02:46.352Z",
-    "modified": "2022-07-04T14:52:54.777Z",
+    "modified": "2022-07-06T11:42:00.079Z",
     "headers": [],
-    "params": [],
+    "params": [
+      {
+        "name": "subsystem",
+        "value": "bankingBot",
+        "isPath": false
+      },
+      {
+        "name": "metric",
+        "value": "logs",
+        "isPath": false
+      }
+    ],
+    "auth": {
+      "type": "bearer",
+      "bearer": "token"
+    },
     "tests": []
   },
   {
@@ -1042,5 +1057,39 @@
       "raw": "{\"hex\":\"04000000000101b7280b37e1a7e66a3828e92137f1f319094fcbf1dd42e833b1deeb9de57a32170100000000ffffffff02081defa30b0000001600147713cba47625c239c09f36254d6bfd37a6b98bb10091949800000000001600141c5b16bc2dc3a9d87bd81b49a0e2f7f4bc6254290002473044022010ed907b9ffc22421a0918033de2e8acfdf1af45ecb549983b98c1bc5c9d79f9022076a307c35ef1a5894eb3502a3d3883e86f609f5e57c87379f6363e3abd39cf510121039a8c6ed7d007b1dbd3199ea1f62fc2c3229e93efaf4cb18200af793155fb4f2500000000\"}",
       "form": []
     }
+  },
+  {
+    "_id": "6c942c09-e7e1-48b1-8534-3f6bfb0ddcb1",
+    "colId": "62eefab7-84cd-496b-8f93-253c15d4f841",
+    "containerId": "fd2e8b01-0e57-4f22-bbb5-226617dc19db",
+    "name": "Post Monitoring Data",
+    "url": "{{url}}/v1/monitoring/data?subsystem=bankingBot&metric=logs",
+    "method": "POST",
+    "sortNum": 340000,
+    "created": "2022-07-05T06:51:11.436Z",
+    "modified": "2022-07-06T11:41:32.959Z",
+    "headers": [],
+    "params": [
+      {
+        "name": "subsystem",
+        "value": "bankingBot",
+        "isPath": false
+      },
+      {
+        "name": "metric",
+        "value": "logs",
+        "isPath": false
+      }
+    ],
+    "body": {
+      "type": "json",
+      "raw": "{\n    \"dfx\": {\n        \"date\": \"today\",\n        \"fileName\": \"testName3\",\n        \"status\": \"ERROR\",\n        \"message\": \"test message 3\"\n    }\n}",
+      "form": []
+    },
+    "auth": {
+      "type": "bearer",
+      "bearer": "token"
+    },
+    "tests": []
   }
 ]
